requirements.txt
# ── Core ML ───────────────────────────────────────────────────
--extra-index-url https://download.pytorch.org/whl/cpu
torch
transformers
sentence-transformers
faiss-cpu
numpy

# ── API & Server ──────────────────────────────────────────────
fastapi
uvicorn[standard]
groq
python-dotenv

# ── Preprocessing (not needed at runtime) ─────────────────────
PyPDF2
nltk
